eval/requirements.txt
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
imageio==2.15.0
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
lpips==0.1.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.64
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-fid==0.2.1
PyWavelets==1.1.1
schedule==1.1.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.63.0
typing_extensions==4.1.1
zipp==3.6.0
